#33 Move array accessor to array module
